requirements.txt
customtkinter
mysql-connector-python

